requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pillow==9.1.1
pycoingecko==2.2.0
pydantic==1.9.1
requests==2.27.1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
yaweather==1.2.2